+++ tests/files/src/0129_suppress_stmts.php
@@ -22,6 +22,7 @@ class C {
      * @suppress PhanUndeclaredVariable
      * @suppress PhanUndeclaredVariable
      * @suppress PhanImpossibleCondition
+     * @suppress PhanImpossibleWeakTypeComparison
      */
     function f() {
         $v = Undef::undef();