+++ tests/files/src/0254_array_bool_comparison.php
@@ -9,7 +9,7 @@ function f1(array $p, array $p2) {
  * @param string[] $p
  * @param string[] $p2
  */
-function f2($p, $p2) {
+function f2($p, array $p2) {
     if($p && true);
     if(true && $p);
     if($p == "string");